Fixes #18 legendTargetFilter now set to tribe_events_month_before_the_grid
Some other minor changes to the public class, too: the teccc_legend hook has been removed (not convinced it is valuable now / in the right place if we do want it for backwards compatibility reasons).

teccc_legend_html filter added to make it easier to customize the legend, wrap it in an extra div etc.